Remove CleanupExtension and retrieve service directly from registry
The reason why the service was passed on through an extension
(of the root project) was to avoid classloader issues. Different
projects might see different versions of the service class.
We now solve this by using reflection.

When we split the build in the future, we will have to use reflection
in any case, because then the extension itself may be loaded in
different classloaders (see #14697).